# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DDAS API",
    "version": "2.0.0-alpha.0",
    "description": "Dynamic Delegation of Authority System — appetite-constrained authority routing for humans and AI agents."
  },
...
  "paths": {
...
    "/api/v1/org/import": {
      "post": {
        "tags": [
          "org"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "units": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "parent": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "key",
                        "name"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "people": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "kind": {
                          "type": "string",
                          "enum": [
                            "human",
                            "agent"
                          ]
                        },
                        "owner": {
                          "type": "string"
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "admin",
                              "policy_author",
                              "approver",
                              "requester",
                              "auditor",
                              "viewer"
                            ]
                          }
                        }
                      },
                      "required": [
                        "key",
                        "name"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "positions": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string"
                        },
                        "unit": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        },
                        "tier": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "holder": {
                          "type": "string"
                        },
                        "validFrom": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "key",
                        "unit",
                        "title",
                        "tier"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "delegations": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "from": {
                          "type": "string"
                        },
                        "to": {
                          "type": "string"
                        },
                        "maxTier": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "scopeUnit": {
                          "type": "string"
                        },
                        "validFrom": {
                          "type": "string"
                        },
                        "validTo": {
                          "type": "string"
                        },
                        "reason": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "from",
                        "to",
                        "maxTier",
                        "validFrom",
                        "reason"
                      ],
                      "additionalProperties": false
                    }
                  }
                },
                "required": [
                  "units",
                  "people",
                  "positions"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "boolean",
              "default": false
            },
            "in": "query",
            "name": "dryRun",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "dryRun": {
                      "type": "boolean"
                    },
                    "units": {
                      "type": "number"
                    },
                    "people": {
                      "type": "number"
                    },
                    "positions": {
                      "type": "number"
                    },
                    "delegations": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "dryRun",
                    "units",
                    "people",
                    "positions",
                    "delegations"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/org/units": {
      "post": {
        "tags": [
          "org"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "parentId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "name"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/org/positions": {
      "post": {
        "tags": [
          "org"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "orgUnitId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "authorityTier": {
                    "type": "integer",
                    "minimum": 0
                  }
                },
                "required": [
                  "orgUnitId",
                  "title",
                  "authorityTier"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/org/position-assignments": {
      "post": {
        "tags": [
          "org"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "positionId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "principalId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "validFrom": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "validTo": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "positionId",
                  "principalId",
                  "validFrom"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/org/delegations": {
      "post": {
        "tags": [
          "org"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fromPrincipalId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "toPrincipalId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "maxTier": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "orgUnitScopeId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "validFrom": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "validTo": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "fromPrincipalId",
                  "toPrincipalId",
                  "maxTier",
                  "validFrom",
                  "reason"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/org/delegations/{id}": {
      "delete": {
        "tags": [
          "org"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}